split views and models
