+++ examples/flaskr/flaskr/auth/views.py
@@ -8,30 +8,9 @@ from flask import render_template
 from flask import request
 from flask import session
 from flask import url_for
-from sqlalchemy.ext.hybrid import hybrid_property
-from werkzeug.security import check_password_hash
-from werkzeug.security import generate_password_hash
 
 from flaskr import db
-
-
-class User(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String, unique=True, nullable=False)
-    _password = db.Column("password", db.String, nullable=False)
-
-    @hybrid_property
-    def password(self):
-        return self._password
-
-    @password.setter
-    def password(self, value):
-        """Store the password as a hash for security."""
-        self._password = generate_password_hash(value)
-
-    def check_password(self, value):
-        return check_password_hash(self.password, value)
-
+from flaskr.auth.models import User
 
 bp = Blueprint("auth", __name__, url_prefix="/auth")
