+++ examples/flaskr/flaskr/blog/views.py
@@ -8,31 +8,8 @@ from flask import url_for
 from werkzeug.exceptions import abort
 
 from flaskr import db
-from flaskr.auth import login_required
-from flaskr.auth import User
-
-
-class Post(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    author_id = db.Column(db.ForeignKey(User.id), nullable=False)
-    created = db.Column(
-        db.DateTime, nullable=False, server_default=db.func.current_timestamp()
-    )
-    title = db.Column(db.String, nullable=False)
-    body = db.Column(db.String, nullable=False)
-
-    # User object backed by author_id
-    # lazy="joined" means the user is returned with the post in one query
-    author = db.relationship(User, lazy="joined", backref="posts")
-
-    @property
-    def update_url(self):
-        return url_for("blog.update", id=self.id)
-
-    @property
-    def delete_url(self):
-        return url_for("blog.delete", id=self.id)
-
+from flaskr.auth.views import login_required
+from flaskr.blog.models import Post
 
 bp = Blueprint("blog", __name__)
